surya: manifest: Remove tetheroffload.config entry
Waited one second for android.hardware.tetheroffload.config@1.0::IOffloadConfig/default

Since android.hardware.tetheroffload.config@1.0::IOffloadConfig/default is not registered, trying to start it as a lazy HAL.

getService: Trying again for android.hardware.tetheroffload.config@1.0::IOffloadConfig/default...

Unable to set property "ctl.interface_start" to "android.hardware.tetheroffload.config@1.0::IOffloadConfig/default": error code: 0x20

Tried to start android.hardware.tetheroffload.config@1.0::IOffloadConfig/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.

Change-Id: I69a9b317ed8ec73e239e5aed7643ab5a146ce3e6